Cleaned up the GUI. Checked to see if there was a solution to CFURLDestroyResource (no). Updated documentation.

+++ BlackRaccoon/BlackRaccoon/BRRequestListDirectory.m
@@ -244,12 +244,12 @@
         }
         break;
             
-        case NSStreamEventEndEncountered: 
+        case NSStreamEventEndEncountered:
         {
             NSUInteger  offset = 0;
             CFIndex     parsedBytes;
             uint8_t *bytes = (uint8_t *)[self.receivedData bytes];
-            int totalbytes = [self.receivedData length];
+            NSUInteger totalbytes = [self.receivedData length];
            
             //----- we have all the data for the directory listing. Now parse it.
             do
